+++ core/routes/callback.py
@@ -38,13 +38,6 @@ class Callback(Resource):
         location="json",
         help="Custom name of the component. It should be a string",
     )
-    # parser_post.add_argument(
-    #     "endpoints",
-    #     type="str",
-    #     required=False,
-    #     location="json",
-    #     help="Endpoints of the component. It should be a dictionary",
-    # )
     parser_post.add_argument(
         "markdown",
         type=str,
@@ -81,9 +74,6 @@ class Callback(Resource):
             custom_name = decode_base64(
                 encoded_data=self.parser_post.parse_args()["custom_name"]
             )
-            # endpoints = self.parser_post.parse_args()["endpoints"]
-            # if endpoints:
-            #     endpoints = decode_base64(encoded_data=endpoints)
             markdown = decode_base64(
                 encoded_data=self.parser_post.parse_args()["markdown"]
             )