+++ CHANGELOG.md
@@ -7,12 +7,21 @@
 - New namespace `influxdb2` for handler InfluxDB requests.
 
 ### Changed
+## [v0.5.2] - 2025-05-16
+
+### Fixed
+
+- Accept `tn_vxlan` and `tn_bastion` as components in the trial network descriptor.
+
+### Removed
+
+- Sites token in base64 format.
+
+## [v0.5.1] - 2025-04-10
 
 ### Fixed
 
 - Bug creating logs when a trial network is validated.
-
-### Removed
 
 ## [v0.5.0] - 2025-03-31
 
@@ -303,7 +312,9 @@
 
 - Frontend implementation.
 
-[Unreleased]: https://github.com/6G-SANDBOX/TNLCM/compare/v0.5.0...HEAD
+[unreleased]: https://github.com/6G-SANDBOX/TNLCM/compare/v0.5.2...HEAD
+[v0.5.2]: https://github.com/6G-SANDBOX/TNLCM/compare/v0.5.1...v0.5.2
+[v0.5.1]: https://github.com/6G-SANDBOX/TNLCM/compare/v0.5.0...v0.5.1
 [v0.5.0]: https://github.com/6G-SANDBOX/TNLCM/compare/v0.4.5...v0.5.0
 [v0.4.5]: https://github.com/6G-SANDBOX/TNLCM/compare/v0.4.4...v0.4.5
 [v0.4.4]: https://github.com/6G-SANDBOX/TNLCM/compare/v0.4.3...v0.4.4